pyproject.toml
[tool.poetry]
name = "inspyollama-client"
version = "0.1.0"
description = ""
authors = ["Taylor B. <tayjaybabee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pywin32 = { version = "*", markers="sys_platform == 'win32'" }
prompt-toolkit = "^3.0.48"
ollama = "^0.3.3"
yaspin = "^3.0.2"
platformdirs = "^4.2.1"
pysimplegui = "^5.0.4"
tabulate = "^0.9.0"
rich = "^13.7.1"
pillow = "^11.0.0"
inspyre-toolbox = "1.6.0-dev13"


[tool.poetry.group.dev.dependencies]
ptipython = "^1.0.1"
ipyboost = {path = "ipyboost-1.0.0a5-py3-none-any.whl"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
